api-template-shared/pyproject.toml
[project]
name = "api_template_shared"
version = "0.0.1"
description = "Shared dependencies for both api and the worker services."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "logfire[redis,httpx,system-metrics]>=3.12.0",
    "loguru>=0.7.3",
    "opentelemetry-distro[otlp]>=0.52b0",
    "opentelemetry-instrumentation-logging>=0.52b0",
    "opentelemetry-instrumentation-redis>=0.52b0",
    "pydantic-settings>=2.8.1",
    "taskiq-aio-pika>=0.4.1",
    "taskiq-redis>=1.0.4",
    "taskiq[opentelemetry]>=0.12.1",
]

[build-system]
requires = ["uv_build>=0.9,<0.14"]
build-backend = "uv_build"
